api/requirements.txt
Django==1.8.1
gunicorn==19.3.0
psycopg2==2.6
djangorestframework==3.1.0
django-rest-swagger==0.3.0
